Introducing configmaps in sample device plugins
Previously configs were directly copied into the the Dockerfile
which meant that anytime configs were updated, the image had to be
rebuilt in order to update the config.

Configmaps gives us more flexibilty and overall a better user experience
so we are switching to using those.

Please NOTE that there is no way in Kubernetes to automatically pick up
updates made to the configmap which means that in case config needs
update a new configmap needs to be created and deployement/ds needs
to be updated to refer to the new configmap. Please refer to the
Kubernetes issue for more issue on this subject here:
https://github.com/kubernetes/kubernetes/issues/22368

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>