devicefile update: Move away from using serial ports as devices
We currently rely on using serial ports as devices. But this can be
limiting because of there being limited serial port available on the
We are forced to keep track of the device count to make sure
that a unique serial port device corresponds to a device exposed
to the kubelet. In addition to that, this puts an implied limit
on the number of devices that can be configured and subsequently
allocated by the sample-device-plugin.

In order to avoid this and prevent any futuure unexpected bugs,
we use `/dev/null` as a device to be allocated from the host to
the container. This gives us flexibility and freedom to emulate
any number of devices. By doing so, the number of devices are
exposed are only dependent on device plugin config and not any
other underlying limiting factor.

This patch also removes references to the devicename in the config
as that is no longer needed.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>