manifests:ci: add kind specific device config
Kind's nodes are single-numa nodes.
Hence, if a pod is asking for a device ,which is configured to be at NUMA node 1, it will fail to be schedule with TopologyAffinityError.
In order to fix the e2e test on kind cluster, we add a specific configuration for the kind-worker node, so the test pod will be able to request (and get) the device.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>