stub: introduce the stub package
This package holds the stub implementation and configuration for the device plugin.

It will allow greater flexibilty, for exmaple controling the device type (it doesn't necessarily has to be /dev/null), resource name, etc.

In addition it exposes the MakePluginApiDevices function which allows building plugin api devices` structure from the sample-device-plugin internal structure

Signed-off-by: Talor Itzhak <titzhak@redhat.com>