manifests: add support for kustomize
Use kustomize in order to deploy sample-device-plugin.
This will allow an easier deployment (all in one keystroke) and more flexible device configuration which is going to be added in a following commit.
We left the old configuration for now, but it can be removed if we decide that we're happy with this change.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>