kustomize: move deviceplugin* under default overlay
We seldom use deviceplugin* overlay directly and for that we also have a dedicated make target anyway.
In order to keep the overlay directory minimal as possible, let's move the deviceplugin* overlays under the default.
This will also allow us to use the deviceplugin* configuration and remove the unnecessary duplication.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>